+++ packages/cli/src/platform/auth/login.ts
@@ -6,15 +6,17 @@ import http from 'http'
 import { green } from 'kleur/colors'
 import open from 'open'
 
-import { name as PRISMA_CLI_NAME, version as PRISMA_CLI_VERSION } from '../../../package.json'
-import { platformConsoleUrl, readAuthConfig, successMessage, writeAuthConfig } from '../platformUtils'
+import { version as PRISMA_CLI_VERSION } from '../../../package.json'
+import { credentialsFile } from '../lib/credentials'
+import { consoleUrl } from '../lib/pdp'
+import { unknownToError } from '../lib/prelude'
+import { successMessage } from '../lib/utils'
 
-interface AuthResult {
+interface CallbackData {
   token: string
   user: {
     id: string
     displayName: string
-    handle: string
     email: string
   }
 }
@@ -27,112 +29,109 @@ export class Login implements Command {
   }
 
   public async parse() {
-    const authConfig = await readAuthConfig()
-    if (isError(authConfig)) throw authConfig
-    if (authConfig.token)
-      return `Already authenticated. Run ${green(
-        getCommandWithExecutor('prisma platform auth show --early-access'),
-      )} to see the current user.`
+    const credentials = await credentialsFile.load()
+    if (isError(credentials)) throw credentials
+    if (credentials) return `Already authenticated. Run ${green(getCommandWithExecutor('prisma platform auth show --early-access'),)} to see the current user.` // prettier-ignore
 
     console.info('Authenticating to Prisma Platform CLI via browser')
 
     const server = http.createServer()
-    // When passing 0 as a port to listen, the OS will assign a random available port
-    const authRedirectUrl = await listen(server, 0, '127.0.0.1')
-    const authSigninUrl = await generateAuthSigninUrl({ connection: `github`, redirectTo: authRedirectUrl.href })
+    /**
+     * When passing 0 as a port to listen, the OS will assign a random available port
+     */
+    const randomPort = 0
+    const redirectUrl = await listen(server, randomPort, '127.0.0.1')
+    const loginUrl = await createLoginUrl({ connection: `github`, redirectTo: redirectUrl.href })
 
     console.info('Visit the following URL in your browser to authenticate:')
-    console.info(link(authSigninUrl.href))
+    console.info(link(loginUrl.href))
 
-    try {
-      const [authResult] = await Promise.all([
-        new Promise<AuthResult>((resolve, reject) => {
-          server.once('request', (req, res) => {
-            server.close()
-            res.setHeader('connection', 'close')
-            const searchParams = new URL(req.url || '/', 'http://localhost').searchParams
-            const token = searchParams.get('token') ?? ''
-            const error = searchParams.get('error')
-            const location = new URL('/auth/cli', platformConsoleUrl)
+    const callbackResult = await Promise.all([
+      new Promise<CallbackData>((resolve, reject) => {
+        server.once('request', (req, res) => {
+          server.close()
+          res.setHeader('connection', 'close')
+          const searchParams = new URL(req.url || '/', 'http://localhost').searchParams
+          const token = searchParams.get('token') ?? ''
+          const error = searchParams.get('error')
+          const location = new URL('/auth/cli', consoleUrl)
 
-            if (error) {
-              location.pathname += '/error'
-              location.searchParams.set('error', error)
-              reject(new Error(error))
+          if (error) {
+            location.pathname += '/error'
+            location.searchParams.set('error', error)
+            reject(new Error(error))
+          } else {
+            // TODO: Consider getting the user via Console API instead of passing it via query params
+            const user = decodeUser(searchParams.get('user') ?? '')
+            if (user) {
+              searchParams.delete('token')
+              searchParams.delete('user')
+              location.pathname += '/success'
+              const nextSearchParams = new URLSearchParams({
+                ...Object.fromEntries(searchParams.entries()),
+                email: user.email,
+              })
+              location.search = nextSearchParams.toString()
+              resolve({ token, user })
             } else {
-              // TODO: Consider getting the user via Console API instead of passing it via query params
-              const user = parseUser(searchParams.get('user') ?? '')
-              if (user) {
-                searchParams.delete('token')
-                searchParams.delete('user')
-                location.pathname += '/success'
-                const nextSearchParams = new URLSearchParams({
-                  ...Object.fromEntries(searchParams.entries()),
-                  email: user.email,
-                })
-                location.search = nextSearchParams.toString()
-                resolve({ token, user })
-              } else {
-                location.pathname += '/error'
-                location.searchParams.set('error', 'Invalid user')
-                reject(new Error('Invalid user'))
-              }
+              location.pathname += '/error'
+              location.searchParams.set('error', 'Invalid user')
+              reject(new Error('Invalid user'))
             }
+          }
 
-            res.statusCode = 302
-            res.setHeader('location', location.href)
-            res.end()
-          })
-          server.once('error', reject)
-        }),
-        open(authSigninUrl.href),
-      ])
+          res.statusCode = 302
+          res.setHeader('location', location.href)
+          res.end()
+        })
+        server.once('error', reject)
+      }),
+      open(loginUrl.href),
+    ])
+      .then((results) => results[0])
+      .catch(unknownToError)
 
-      const result = await writeAuthConfig({ token: authResult.token })
-      if (isError(result)) throw result
+    if (isError(callbackResult)) throw new Error(`Authentication failed: ${callbackResult.message}`) // prettier-ignore
 
-      return successMessage(`Authentication successful for ${authResult.user.email}`)
-    } catch (error) {
-      throw new Error(`Authentication failed: ${isError(error) ? error.message : ''}`)
+    {
+      const writeResult = await credentialsFile.save({ token: callbackResult.token })
+      if (isError(writeResult)) throw new Error('Writing credentials to disk failed', { cause: writeResult })
     }
+
+    return successMessage(`Authentication successful for ${callbackResult.user.email}`)
   }
 }
 
-const generateAuthSigninUrl = async (params: { connection: string; redirectTo: string }) => {
-  const cliSignature = await checkpoint.getSignature().catch((e) => {
-    debug(`await checkpoint.getSignature() failed silently with ${e}`)
-    return null
-  })
+const createLoginUrl = async (params: { connection: string; redirectTo: string }) => {
+  const signature = await checkpoint.getSignature().catch(unknownToError)
+  if (isError(signature)) debug(`await checkpoint.getSignature() failed silently with ${signature.message}`) // prettier-ignore
 
-  // todo change how we pass signature. Put into header user-agent, format idea: <sig>.<ver>.client.prisma
-  const state = {
-    client: `${PRISMA_CLI_NAME}@${PRISMA_CLI_VERSION}`,
-    // will be `null` if it throws during retrieval
-    // will be a UUIDv4 when successful
-    signature: cliSignature,
+  const state: State = {
+    client: `${isError(signature) ? 'error' : signature}.${PRISMA_CLI_VERSION}.cli.prisma`,
     ...params,
   }
-  const stateEncoded = Buffer.from(JSON.stringify(state), `utf-8`).toString(`base64`)
-  const url = new URL('/auth/cli', platformConsoleUrl)
+  const stateEncoded = encodeState(state)
+  const url = new URL('/auth/cli', consoleUrl)
   url.searchParams.set('state', stateEncoded)
   return url
 }
-
-const isConsoleUser = (maybeUser: unknown): maybeUser is AuthResult['user'] => {
-  if (typeof maybeUser !== 'object' || maybeUser === null) return false
-  const user = maybeUser as AuthResult['user']
-  return (
-    typeof user.id === 'string' &&
-    typeof user.displayName === 'string' &&
-    typeof user.handle === 'string' &&
-    typeof user.email === 'string'
-  )
+interface State {
+  client: string
+  connection: string
+  redirectTo: string
 }
 
-const parseUser = (stringifiedUser: string) => {
+const encodeState = (state: State) => Buffer.from(JSON.stringify(state), `utf-8`).toString(`base64`)
+
+const decodeUser = (stringifiedUser: string) => {
   try {
     const maybeUser = JSON.parse(Buffer.from(stringifiedUser, `base64`).toString(`utf-8`))
-    return isConsoleUser(maybeUser) ? maybeUser : null
+    if (typeof maybeUser !== 'object' || maybeUser === null) return false
+    const isUser =
+      typeof maybeUser.id === 'string' &&
+      typeof maybeUser.displayName === 'string' &&
+      typeof maybeUser.email === 'string'
+    return isUser ? maybeUser : null
   } catch (e) {
     debug(`parseUser() failed silently with ${e}`)
     return null
