+++ packages/cli/src/platform/auth/show.ts
@@ -1,7 +1,9 @@
 import { Command, formatTable } from '@prisma/internals'
 import { green } from 'kleur/colors'
 
-import { argOrThrow, getPlatformTokenOrThrow, platformParameters, platformRequestOrThrow } from '../platformUtils'
+import { argOrThrow } from '../lib/parameters'
+import { requestOrThrow } from '../lib/pdp'
+import { getTokenOrThrow, platformParameters } from '../lib/utils'
 
 export class Show implements Command {
   public static new(): Show {
@@ -10,8 +12,8 @@ export class Show implements Command {
 
   public async parse(argv: string[]) {
     const args = argOrThrow(argv, { ...platformParameters.global })
-    const token = await getPlatformTokenOrThrow(args)
-    const { me } = await platformRequestOrThrow<{
+    const token = await getTokenOrThrow(args)
+    const { me } = await requestOrThrow<{
       me: {
         user: {
           id: string
