+++ packages/cli/src/platform/lib/pdp.ts
@@ -1,0 +1,68 @@
+import fetch, { Headers } from 'node-fetch'
+
+import { name as PRISMA_CLI_NAME, version as PRISMA_CLI_VERSION } from '../../../package.json'
+
+const platformAPIEndpoint = new URL('http://localhost:8788/api')
+
+export const consoleUrl = new URL('https://console.prisma.io')
+
+/**
+ *
+ * @remarks
+ *
+ * TODO Feedback from Joel:
+ *    It could be interesting to set a default timeout because it's not part of fetch spec, see:
+ *    npmjs.com/package/node-fetch#request-cancellation-with-abortsignal
+ */
+export const requestOrThrow = async <
+  $Data extends object = object,
+  $Input extends null | object = null,
+  $Variables extends null | object = null,
+>(params: {
+  token: string
+  body: $Input extends null
+    ? $Variables extends null
+      ? {
+          query: string
+          variables?: undefined
+        }
+      : {
+          query: string
+          variables: $Variables
+        }
+    : {
+        query: string
+        variables: { input: $Input }
+      }
+}): Promise<$Data> => {
+  const method = 'POST'
+  const headers = new Headers({
+    'Content-Type': 'application/json',
+    Authorization: `Bearer ${params.token}`,
+    // See https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/User-Agent
+    'User-Agent': `${PRISMA_CLI_NAME}/${PRISMA_CLI_VERSION}`,
+  })
+  const body = JSON.stringify(params.body)
+  const response = await fetch(platformAPIEndpoint.href, { method, headers, body })
+  const text = await response.text()
+  if (response.status >= 400) throw new Error(text)
+
+  const json = JSON.parse(text) as { data: $Data; error: null | object }
+  if (json.error) throw new Error(`Error from PDP Platform API: ${text}`)
+
+  const error = Object.values(json).filter(
+    (_): _ is { __typename: string } => typeof _ === 'object' && _ !== null && _['__typename']?.startsWith('Error'),
+  )[0]
+  if (error) throw errorFromPlatformError({ message: '<message not selected from server>', ...error })
+
+  return json.data
+}
+
+export const errorFromPlatformError = (error: PlatformError): Error => {
+  return new Error(error.message)
+}
+
+export type PlatformError<$Types extends string = string> = {
+  __typename: $Types
+  message: string
+}
