+++ packages/cli/src/platform/lib/prelude.ts
@@ -1,0 +1,4 @@
+export const unknownToError = (unknown: unknown): Error => {
+  if (unknown instanceof Error) return unknown
+  return new Error(`Unknown error: ${unknown}`)
+}
